requirements.txt
flask>=3.0
flask-restx>=1.3.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
chromadb>=0.5.0
python-dotenv>=1.0.0
tenacity>=8.0.0
google-generativeai>=0.8.0
